+++ provision/base.sh
@@ -192,8 +192,7 @@ EO_MAKE_CONF
 }
 
 configure_fstab() {
-	local _sub_dir=${1:-""}
-	local _etc_path="$BASE_MNT/${_sub_dir}etc"
+	local _etc_path="$BASE_MNT/etc"
 	if [ ! -d "$_etc_path" ]; then mkdir -p "$_etc_path"; fi
 
 	store_config "$_etc_path/fstab" "overwrite" <<EO_FSTAB
@@ -234,8 +233,6 @@ configure_base()
 	configure_bourne_shell "$BASE_MNT"
 	configure_csh_shell "$BASE_MNT"
 	touch "$BASE_MNT/etc/fstab"
-	configure_fstab "data/"
-	install_pfrule base
 }
 
 install_periodic_conf()